manage_restaurant_api/requirements.txt
amqp==5.3.1
asgiref==3.9.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.4
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
kombu==5.5.4
packaging==25.0
phonenumbers==9.0.8
prompt_toolkit==3.0.51
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
